parser_common: eliminates the repetition of common code between vinumc and vin2dot This will be useful for maintaining consistency in error messages.
